Uses Wireguard for tunnels
First draft, has a bunch of limitations like 80/443 hardcoded, and no
state updating (if ports change, nothing else does).

+++ innisfree/ssh.py
@@ -1,13 +1,14 @@
-import tempfile
 import subprocess
 import os
 from ruamel.yaml.scalarstring import PreservedScalarString as pss
 from typing import Tuple
 from pathlib import Path
+from .utils import CONFIG_DIR
 
 
 class SSHKeypair:
-    def __init__(self) -> None:
+    def __init__(self, prefix="") -> None:
+        self.prefix = prefix
         (privkey_filepath, privkey, pubkey) = self._create()
         self.filepath = privkey_filepath
         self.private = privkey
@@ -19,10 +20,13 @@ class SSHKeypair:
         Uses ssh-keygen to generate, since SSH uses a special format. More research
         required to use e.g. cryptography to generate in Python without shelling out.
         """
-        _, privkey_filepath = tempfile.mkstemp()
-        pubkey_filepath = privkey_filepath + ".pub"
+        privkey_filepath = CONFIG_DIR.joinpath(f"{self.prefix}id_ed25519")
+        pubkey_filepath = Path(str(privkey_filepath) + ".pub")
         # Deleting because ssh-keygen can't clobber files
-        os.unlink(privkey_filepath)
+        if os.path.exists(privkey_filepath):
+            os.unlink(privkey_filepath)
+        if os.path.exists(pubkey_filepath):
+            os.unlink(pubkey_filepath)
 
         cmd = [
             "ssh-keygen",
@@ -43,7 +47,7 @@ class SSHKeypair:
         with open(pubkey_filepath, "r") as f:
             pubkey = f.read().rstrip()
 
-        return (Path(privkey_filepath), privkey, pubkey)
+        return (privkey_filepath, privkey, pubkey)
 
     def __repr__(self) -> str:
         s = f"<SSHKeypair pubkey: {self.public}, filepath: {self.filepath}"
